Disable shellcheck test for iterating over ls output
Disables:
SC2045: Iterating over ls output is fragile. Use globs.
https://github.com/koalaman/shellcheck/wiki/SC2045

The alternative is more verbose particularly since the filename also
has to be extracted for use later in the function. Since the filenames
won't contain the types of special characters that make iterating
ls output risky, it's fine to just disable the check.